Enable the performace CPU governor on Hadoop workers
This will cause the Hadoop workers to use the performance governor
instead of the ondeman governor, which we believe will boost the
performance of the Hadoop nodes.

There will be a moderate increase in power usage across the Hadoop cluster.

Owing to the fact that eqiad rack B4 is already alerting for high power
usage, the five hosts in this rack have been excluded from the change.

Hosts: O:analytics_cluster::hadoop::worker
Hosts: an-worker1085.eqiad.wmnet
Bug: T365878
Bug: T362922
Change-Id: I511c19ccbcc72aaf134e4d056f7b8a5cbd2ad155