Add additional ignore line to Juniper warnings for Homer
Adds additional line to warning messages we can ignore when running
Homer.  This one occurs on SRX firewall clusters when a 'replace'
is done on the interface section.  Occurs even when there is no
diff, and tested with and without diffs and in neither case was
traffic affected or anything changed.

Bug: T378070
Change-Id: Ic9ab0794357110650f0eb6654c4ee6114f9b0ecf