services_proxy: add tcp_keepalive parameter, enable for shellbox-video
Rather than add tcp timeouts to shellbox itself, have envoy add a tcp
timeout to the connection so as to make LVS happy when using long (>
20m) connections. Without this, LVS will lose track of our connection
and sent an RST when the server starts to respond.

This option uses the config.core.v3.TcpKeepalive semantics, and some
or all parameters can be set as needed.

Bug: T373517
Change-Id: I9b9ac2bb9604ecbf8315602c7fc6c38c615bb149