P:conftool::client: make conftool2git_host Optional
A default_value of empty string is incompatible with Stdlib::Fqdn, which
breaks compilation when the hiera key is unset.

While this clearly isn't a problem in production, it does lead to some
PCC noise from cloud hosts that don't (and shouldn't) have this set.

Instead, switch to Optional[Stdlib::Fqdn] with default_value of undef.

Hosts: P:conftool::client
Change-Id: I55d8c665e627f798d7676a5478755339bbc03902