durum: ensure nftables DENY sets are present (does not mean drop)
This ensures the nftables DENY sets and throttle rule get created
but when the policy is still set to accept it will never mean
any packets get dropped.

It just means we start tracking IPs that would trigger a treshold.

And also it will fix the issue that the prometheus exporter service
fails because it can't find data.

follow-up to Ie5f985203af9

Change-Id: I0947830aa9d43ba44302886ca5e9925fed673c81