P:cache::varnish::frontend: add check for minimum frontend cache size
From the learnings in T378724, some cp hosts in magru ended up with the
following configuration:

    sukhe@cp7001:~$ free -h
                   total
    Mem:           125Gi

This then resulted in the following varnish configuration:

    -s malloc,1G -s Transient=malloc,5G

When in reality it should have been:

    -s malloc,334G -s Transient=malloc,10G

Assuming that the host had the right amount of physical memory of
course, which was not the case here. But varnish was happy to allocate
1GB of malloc storage and we would have never know about this.

To fix this, we add an optional boolean to warn us if the amount of
frontend memory reserved is less than a magic number of 150GB, that is
computed based on the current hardware configuration. If there is
anything less than that, Puppet compilation/initial reimaging should
fail and warn us about some possible incorrect hardware configuration,
at which point it is for the operator to figure out what is happening.

Bug: T378724
Hosts: cp4037.ulsfo.wmnet, cp1100.eqiad.wmnet
Change-Id: I594ef509f025430f2d4444be6f3cc45f532f2190