profile::docker::report: use the internal registry endpoint
We used to set (IIUC) the external docker endpoint only with
the hope to eventually leverage better caching from the CDN.
So far we don't have any specific caching strategy set
for docker-registy.wikimedia.org, beside the default (1 day max for objs)
that would probably not be effective when running docker
report every week (the biggest job for all the service-related
images, not the base ones).

Set the internal endpoint and remove any proxy setting.

Bug: T378618
Change-Id: I9df40dbe9188c866c2e7e9b885607c30d9ba9118