docker_registry_ha: reduce from 300 to 180 the nginx timeout
We used to keep increasing the nginx proxy timeouts for the Registry,
with the hope to allow the catalog calls to finish without ending up
in getting closed by nginx because too much time elapsed.

We realized in T376285 that it was all due to the registry
being rate limited by Swift, and once we lifted the limits
the catalog calls started to complete in a reasonable amount of
time.

Set a more reasonable value for timeouts, 180s is in-line with
the ATS/CDN ttfb timeout and it is more reasonable than 300s.

Bug: T378618
Change-Id: Id75c6f3de1c789146376b8a5e80cf7e4ab38f2ff