Create new lvs service kartotherian-k8s-ssl
In order to be able to support the migration to k8s, a higher
port than 443 needs to be used for kartotherian-ssl. On maps
bare metal nodes there are already multiple listen statements
that support both 443 and 6543, so endpoints like maps.wikimedia.org
at the CDN level will not be impacted by the transition.

This change adds a new endpoint for port 6543, using the same VIP
as the other kartotherian services. Once everything is migrated
over the new ip:port combination, I'll deprecate the :443 one.

Bug: T378944
Change-Id: I626979b7b43383c30c843c2a1b6df3a0fb84e930