backup: add /var/lib/gerrit to gerrit repodata backup filesets
Previously the gerrit repo dir was /var/lib/gerrit2 but on future
servers we are changing the system user from "gerrit2" to "gerrit" and
the repo dir is built from that.

So one directory will only exist on old servers and the other
only on new servers.

It seems easiest to just add both here unless it's an issue for backups
if a directory doesn't exist and it won't simply ignore or skip it if
that's the case.

If that's an issue we would have to create a separate fileset and use
that and add code to use the right one depending which machine it is.

Bug: T338470
Change-Id: I935460d0a8ec6c0e2605b3b6be6b197a31040eaf