gerrit: add chown parameter to lfs data rsync, ensure daemon_user is used
Due to T338470 we will have a situation where the user owning the
rsynced lfs data is different on existing vs new servers.

So currently the permissions are wrong after rsync to a new (not in
production yet) server.

But luckily rsync::quickdatacopy nowadays has the chown parameter
and if we simply set it to $daemon_user this should ensure it's
always using the correct user on the target.

Bug: T338470
Bug: T372804
Change-Id: Ic79ea9d5a35303357ce6285324fff4cca1078aaa