Remove pfw3-eqiad and replace with pfw1-eqiad
The new eqiad pfw cluster, pfw1-eqiad, will not be  reachable until
the uplinks from current pfw3-eqiad to CRs are moved to it next
Tuesday, November 12th.

So effectively it will not be reachable on other than the mgmt
hostname before then, and likewise the old pfw cluster will
become unreachable in-band when they are moved.

This should not therefore be merged until during the works.

Bug: T377381
Change-Id: I335a324b35308f8a62362bdf06364de81ccff4a5