profile::docker::reporter::report: use internal registry endpoint
This is the second attempt to do it. The first one included
the removal of the proxy settings, but it turned out to be a problem
since stuff like security.debian.org couldn't be reached anymore
from build2001.

Set the endpoint in the systemd timer only to the registry's
discovery endpoint.

Known side-effect: on debmonitor, the new reported images will
have, as full name, .discovery.wmnet and not .wikimedia.org

Bug: T378618
Change-Id: I843e20b13f04bba7e8eddbe700c3724dcac28588