Move kartotherian-k8s-ssl LVS endpoint to "production" state
Fix also the current lvs pools for bare-metal maps nodes.
All the kartotherian* LVS endpoints share the same LVS IP,
so the right loopback settings are already deployed on all
the nodes, but since kartotherian-k8s-ssl should become
the only active/maintained one, fix it for consistency
on the maps nodes as well.

Bug: T378944
Change-Id: I5bb97b7d957acedb70813688743f5e746a4b958a