Revise Envoy firewall options
Even before we introduced nftables support the interface was brittle
since the difference between unset and an empty list was easy to
mistake. And the additional complexity of supporting typed arguments
doesn't make it less simple...

To make this all less disambigious I'm introducing an explicit option
to select global access to the Envoy port while using firewall::service.
This takes precendence over all the other options.

Otherwise, if either firewall_srange and/or firewall_src_sets is set,
the access to the Envoy port is restricted based on these parameters.

If neither are set, it falls back to using ferm syntax.

This patch also migrates the existing configs over to use global
access. Some of these were actually mis-configured (the caches
restrictions for phab2002, etherpad, miscweb and planet are currently
not live due to a bug), so this converts the status quo, not what was
meant to be configured.

The actual migration to use the intended restrictions to caches can
happen in followup patches.

In addition three roles had config setting for Envoy, but are not yet
using nftables. I've remove these for now (idp-test, matomo and idm),
this will be re-added in the new config format once this patch has been
merged.

Hosts: C:profile::tlsproxy::envoy
Change-Id: I01ef8b8d5fad6634fa5d55f60a988aada0911928