Add missing params to enable druid test to delete unused segments
We previously added a parameter to the Druid test cluster that enables
the deleting of unused segments. Upon restarting the service on the
druid test cluster we observed a failure to start the coordinator
service.

The documentation suggests that these values have default values, but
our slightly older version must not have these defaults compiled in.

We configure each task with a maximum of 100 segments that it can delete
and we configure a minimum retention time of 90 days.

Hosts: an-test-druid1001.eqiad.wmnet
Bug: T376118
Change-Id: If467187be6e029ed09ce8eb8f5d671ca1e4715bb