resolvconf: don't update resolv.conf with 0 nameservers
We just saw a case on cloud-vps where a lookup for the nameserver failed
but puppet went ahead and updated resolv.conf without a nameserver.

Needless to say, this was bad! It broke lookups on the host which meant
that all future puppet runs failed so the error could not correct itself.

I can't think of why we would ever want a resolv.conf without any
nameservers, so now we never do that.

Hosts: dns1004.wikimedia.org, acmechief1002.eqiad.wmnet
Bug: T379927
Change-Id: Ia3ff61731c194ab453098033ef4ed6ba23a6b100