keepalived: Split failover config template to new class
The current keepalived class has functionality to override the default
config template by passing a config parameter, but that still requires
the parameters to generate that template. Fix that by splitting the code
to use that template to a new class which leaves the base keepalived
class to only manage the package and the service.

Bug: T380057
Change-Id: Idc5be68a643c2a907e23dc9789c81b6cc58dba1d