Remove profile::ldap::bitu from Cumin role
This was once added in 239915883a76354d8c720e93e0ec40f9ea32399e with
the intent to be used in an offboarding cookbook, but that never
happened (no cookbook uses it). As such, let's remove the profile
until it's actually needed (and we can stop building bitu-ldap for
Bullseye)

python3-bitu-ldap will be cleaned up manually with Cumin.

Change-Id: I4bbbd5e8a1debba3797981f4d9918c54460f5a17