scap target: ensure scap is installed on host before it is required
Before, there was no order specified for the scap installation. This
was causing scap to be missing when it was required during the
imaging of a new host.

As a further improvement, also make sure the `scap` user creation
happens at the same time scap itself is installed

Bug: T378769
Hosts: R:scap::target
Change-Id: Ia27a497fbde6f01c685359dbb582e14d709954c6