wikikube: Add wikikube-ctrl1004
Why:
wikikube-ctrl1001 is from an old batch and needs to be decommissioned.
To do that, we need to use one of the machines from the refreshing
batch. However all of those have already been imaged as wikikube-worker
nodes, so we need to reimage one of the wikikube-worker nodes. Taking
into account that we need a 10G card and a 10G capable switch with free
ports, wikikube-worker1290 was chosen.

What:
Add wikikube-ctrl1004 stanzas

Hosts: auto
Bug: T379790
Change-Id: Ia7cf8ecd21bbd979027b2fe6e800d7622d3ec1de