wmflib: add facter data for lshw -class memory
The current system memory information returned by facter (read from
/proc/meminfo) can vary across systems, even with the same amount of
total physical memory. To avoid such discrepancy within the same class
of hosts, we want to return the output of lshw -class memory and have it
available in the facter output of hosts.

The primary goal for this is T378724: as an example, two DNS hosts with
the same amount of physical memory (32 GiB) should report it as such.

We confine this to only Traffic hosts for now.

Bug: T378724
Change-Id: Id5fe57103c786528feb69a6850901b5829d450c2