profile::auto_restarts::service: make restart time configurable
This change makes the restart time for the wmf_auto_restart job
configurable using optional $restart_hour and $restart_minute parameters.
If no value is set, the default random calculation is used.

This is needed because in rare cases the random time conflicts with
other jobs running on the host (for example the GitLab replica
restore). So it would be nice to optionally set the timer explicitly.

Bug: T380476
Change-Id: I24c8ab2272339b6fbc86c78ae534d010517e7ae0