Deprecate system::role for Openstack roles
system::role is a bit of a hack. In 21fa33b455bfd956f45525b7bb3c22c77ff3787d
support was added to move the functionality into profile::base::production
and simply set the role description via a Hiera flag. Apply this for the
cloudrabbit role, the others only used the class title, so system::role
can simply be removed for them entirely

Change-Id: Ieb6127beff8862b36c2f98f7cfa5993332ea2545