mailman: Change task runner to operate every 12 hours
We think there's a race condition in the Mailman task runner which is
causing the content of some mails to disappear. There's a few notes in
the release notes for the latest mailman version which mention fixing
one, and this seems to be the likely culprit if that's the case.
Switching the task runner should reduce the frequency of how often we
see this (since its seems to be happening roughly hourly)

Bug: T377045
Change-Id: Iaa3c231678bdaad7eba12e55f637f45c77cf0fb8