scap: delete wmf branches automatically
We used to delete the MediaWiki wmf branches automatically. That was
disabled behind the feature flag `--delete-gerrit-branch`. Since then,
we haven't deleted branches.

I eventually fixed the root cause in T218783 and I have confirmed the
branches are now properly deleted using:

    scap clean --delete-gerrit-branch 1.44.0-wmf.2

The timer kicks in on Monday at 21:00 PST, it would try to delete the
branch and that will pass. Next week it will delete the 1.44.0-wmf.3
branch and if all is fine I can claim victory.

Bug: T303828
Change-Id: I2d2f836abe2e2c0b9262bf716081709c1ebada23