vrts: Update mail alias generation script to bail on too many changes
We had an incident recently where we stopped generating aliases for
every vrts email address in the wikimedia.org domain.

This change limits the number of aliases allowed to be generated at
once, so that if a large batch of changes come in at once (indicating
something might be broken), the script won't run without a --force flag.

Bug: T380009
Change-Id: Ifc549a88a0180689b67dfb07f678b22aac7b53b9