ml-lab: Allow users to run nvtop and radeontop via sudo
radeontop via sudo has a precedent in the statboxes.

While rootless use of nvtop allows to see that a GPU is being used, it
will not show who is using it. Since the usecase of the lab machines is
collaborative use, it's im portant to be able to see who is using the
GPU(s), so users can contact each other for coordination.

nvtop does not allow one to spawn a shell, so it's not an avenue for
free-for-all root escalation. It does allow one to kill someone else's
GPU-using processes, but since this is a pretty small and closed user
group (<20 people, all of them fulltimers or contract employees), it is
probably fine. If not, we can always revoke the privilege or limit it to
the ml-team users on the machine.

Change-Id: I77051a4a11d5e34faf1db0a1e97e9fce9633e6b9