Change IP for lvs7003 on public1-b3-magru to 195.200.68.5/27
lvs7003 vlan711 interface was previously using 195.200.68.4/27,
however when it was decom'ed in Netbox that reservation was removed,
and due to sequencing of re-provisioning of servers that IP got
picked and allocated to dns7001 when it was set up anew.

I've reserved 195.200.68.5/27 for lvs7003 to use in Netbox now
instead.  This patch should change it on the host to remove the IP
conflict and allow both servers to work again.

Bug: T376737
Change-Id: Idd88c3020e0b3d06a65f8d26780b1d86eb86a69b