scap target: ensure scap is installed on host before it is required
Before, there was no order specified for the scap installation. This
was causing scap to be missing when it was required during the
imaging of a new host.

Follow-up to Ia27a497fbde6f01c685359dbb582e14d709954c6 but without
https://gerrit.wikimedia.org/r/c/operations/puppet/+/1092841/5/modules/scap/manifests/init.pp#20
to see if it has an effect on the unexpected user creation cycle:
https://phabricator.wikimedia.org/T378769#10363614

Bug: T378769
Hosts: R:scap::target
Change-Id: Id300efa1a2aa74870e9811ddd61455b87780a7ab