mwdebug: Enable retries
Why:
In T380598, we have witnessed some difficult to reproduce consistently
transient errors that apparently stem from Apache closing the HTTP
connection envoy has to it while envoy tries to send a new request. An
easy way to solve the problem is probably to just retry the request.
Since mwdebug are slated to be decommissioned and will never be upgraded
to PHP 8.1 (since this is going to only happen in mw-on-k8s), this seems
like enough effort to solve a large enough part of the problem described

What:
Switch profile::tlsproxy::envoy::retries to true

Hosts: mwdebug1001.eqiad.wmnet
Bug: T380598
Change-Id: I543a22a5b3560468b5e6af38a12c19bd91f5c7aa