Block PAWS workers nodes from all UDP traffic other than DNS & NTP
These nodes have been sending high volumes of UDP floods (looks to be
denial-of-service) to a wide variety of internet IPs for the past
few weeks.

Blocking for now in this crude way until we can get firmer control on
the situation.  Also NATing them to a different outside IP so we
can easier identify the traffic from them on the other side.

Bug: T381078
Change-Id: Ic3b0598abb42f771b83fa679c3e6e8759058759b