trafficserver: switch query-scholarly to wikikube
This change switches the low-traffic query-scholarly endpoint to
wikikube. It reuses the existing miscweb service. Only the gui is switched,
/sparql and /bigdata endpoints remain on wdqs nodes.
If this test is successful, the other query guis can be switched
as well.

This is the second try of I03e1484e55ca767bb340979a2aba2100b3e06282.
The first version was reverted during unrelated commons-query issues.

Bug: T350793
Change-Id: Ic67e16343ecb4deb58e9ba2019af0468bf99e13a