Increase the number of gnmic worker and writer threads
We have occasional gaps in the gnmic stats when graphed in grafana.
I'm not 100% what the issue is here other than there is likely
a bottleneck in the pipeline somewhere.

Looking at the netflow VMs they are fine in terms of average CPU,
and each have 4 vCPUs.  As such I think it makes sense to set the
reader and writer threads to match the number of available CPUs and
hopefully spread the load a bit.  With any luck it may improve the
situation with the anomolies in graphs.

Bug: T369384
Change-Id: Ic39f9eb9a33bcd9b716d06e955d42608d374a315