[dumps] Increase the lbzip2 thread count for large wikis
When we run the recombine dumps tasks the utility lbzip2 is used to
compress files in parallel to the bzip2 format.

We configure the number of parallel threads in order to find a balance
between the hardware and the load.

We have not increased the parallelism for over six years, even though
the hardware has been upgraded.

The values were originally set in a conservative way in: The Change Id:
Ia81702b99fd9258d920aed97ac93ddda10dabd30

We have also monitored the performance of the hardware and noted that
the system load is low during these recombine operations, so we can
increase it.

Hosts: O:dumps::generation::worker::dumper
Bug: T380729
Change-Id: Ifddd6faa341f2f57b9728711da85d8ca040e9506