Install updated ruby-sys-filesystem on bullseye systems running Puppet 7
The default version of facter in Debian is 3.14.12. The 3.x branch is
written in C++. For the 4.x branch Puppet labs switched back to Ruby
and to parse sysfs structures for the "mountpoints" fact (which was
implemented natively in C++ for 3.x) is uses ruby-sys-filesystem.

We need to install ruby-sys-filesystem for bullseye hosts on Puppet 7
and also switch to a backport since the version in vanilla Bullseye
is too old.

Hosts: sretest1001.eqiad.wmnet, conf1007.eqiad.wmnet,
Hosts: sretest1002.eqiad.wmnet, maps1008.eqiad.wmnet
Bug: T381538
Change-Id: I9ea1de3d7a31864ad532ddd93d8439ff5b8b88b7