P:idm disable index listings for Bitu media and static content.
Pentesters doesn't like directory listings. Bitu does not allow users to
upload content and only serves static content available in the Git
repository, this there makes no difference in terms of actual security.

It is a valid observations for other systems and we should follow best
practises, so disable the listings in case the Puppet code is used as
inspiration for other projects.

Bug: T381637
Change-Id: I1a1e2621842d9df6f4356dd4ba2638bb63860899