Rename kubernetes20[36-39] to wikikube-worker20(47|66|85|86)
Puppet configuration for renaming of kubernetes2036, kubernetes2037,
kubernetes2038 and and kubernetes2039 to wikikube-worker2047,
wikikube-worker2066, wikikube-worker2085 and wikikube-worker2086.

The wikikube-worker20(47|66|85|86) names might conflict with the
ongoing decommissioning in T379788 so I guess we want to skip those names.
But let's discuss this in Gerrit, I can amend a chance with
higher wikikube-worker names.

Bug: T379788
Bug: T377877
Change-Id: I188d1558ca008a3c92df947a8740c833e9145bce