s.m.a.r.t. - Exclude zram devices from data export
We recently enabled a zram device on one of our hosts, but this caused
an alert because /dev/zram0 was being included in the `smart_data_dump`
prometheus metrics.

In order to fix this, we filter out devices named zram* from the json
response of the lsblk command.

Hosts: O:statistics::explorer
Bug: T380835
Change-Id: I59e4a651f6d6325b6604c4d596936cdc33d7e8f4