Move some cephosd hieradata into profile default files
Until now, we have only had to deploy ceph configuration to the ceph
server roles themselves (or into k8s via configmaps).

We are now at the point where we would like to deploy a minimal Ceph
configuration and cephx keyrings to a number of clients, such as the
stat and ml-lab servers.

In order to avoid duplicating the hieradata, we move the values into
profile defaults according to the data centre. This will also allow us
to bring the codfw ceph cluster online more easily.

This change is intended to be a noop.

Hosts: O:ceph::server
Bug: T378735
Change-Id: Id459bd3d75b8b39e8eaf441bcb5354d31d3b79ef