cephosd: enable the deployment of client cephx keys and minimal ceph.conf
We have a mechanism for distributing cephx keys to ceph clients, as well
as defining them on the servers. This is by the use of the
ceph::auth::deploy class, which has not been used so far on the cephosd
cluster.

This commit creates a new class called profile::ceph::client which
generates a minimal ceph config and distributes only those keys which
should be on the server.

By itself, this change should do nothing. It will be followed by a
change to deploy a single key and minimal config to the ml-lab servers.

Hosts: O:ceph::server
Bug: T378735
Change-Id: I5b7e1fda66555576822eb5505c612f718ff64eb4