icinga: fix icinga-status bug
Why:
To ensure that the icinga-status script has read the Icinga's status.dat
file completely and not an half-written version of it, it checks for
elements at the end of the file.

What:
Expand the list of end-of-file elements checked as if there are no
downtimes it will consider the file corrupted.
Makes also the error message more explicit.

Hosts: auto
Change-Id: I904bcaac455ccd2c439a0672f3377ee5eb9fd3bf