Deprecate system::role for phab/mariadb
system::role is a bit of a hack. In 21fa33b455bfd956f45525b7bb3c22c77ff3787d
support was added to move the functionality into profile::base::production
and simply set the role description via a Hiera flag.

We can't easily pass the shared to the Hiera setting, so hardcode it to
the shard it has always been on.

Change-Id: I89be2fec70ef4b0f12f4f55d4dd940147dad6531