Add an-worker106[5-9] to puppet
We are re-purposing five old Presto nodes as Hadoop workers,
temporarily. In order to this, they are being given out-of-sequence
hostnames, an-worker106[5-9].

This commit adds them to site.pp and preseed.yaml and common.yaml.

In addition, we configure an-worker1069 not to run the performance CPU
governor, because it is in rack A4 and all of the other servers in this
rack have also been excluded from this configuration.

Bug: T382410
Change-Id: Ie46d2a68a28eaab9c4b7bd146b1585f46f6af9fe