dse-k8s: Add tokens for mediawiki-dumps-legacy namespace
We are going to be migrating the current generation of dumps to the
dse-k8s cluster, so we need a namespace with matching tokens to begin
working with it.

We might want a specific access group to have deployment rights, but for
now we can use the existing analytics-deployers group to restrict
access.

Hosts: O:deployment_server::kubernetes
Bug: T382489
Change-Id: I24d4cc75b07f741de600750a3557ca99a1b985ec